Prepare archivos para el registros de casos epidemilogicos Borre archivos innecesarios mejore los nombre de los script que limitan las query. Anadi mesaje de caragando y cancelar para ajax, en alerts Termine la seccion de actualizacion del Catalogo CIE-10 Anadi la libreria spout para el manejo de documentos Anadi seccion para listar el Catalogo CIE-10 en la Base de datos

+++ ajax/casosEpidemiAjax.php
@@ -1,29 +1,29 @@
 <?php 
 
-var_dump($_POST);
-
 	$requestAjax = TRUE;
 
 	require_once "../config/app.php";
-	
-	if (isset($_POST['docIdentidad'])) {
-
-		require_once "../controller/personaController.php";
 
 		require_once "../controller/casoEpidemiologicoController.php";
 
+		$casoEpidemiologicoController = new casoEpidemiologicoController();
+	
+if (isset($_POST['operationType']) && $_POST['operationType'] == "register"){
 
+	 	$casoEpidemiologicoController->addCasoEpidemiologicoController($_POST);
+
+}elseif ($_POST['operationType'] === "delete") {
+	 	$casoEpidemiologicoController->deleteCasoEpidemiologicoController($_POST);
+	
 		//$personaController = new personaController();
 	 	//$personaController->getPersonaController($_POST);
-echo "<br><br>";		
-		$casoEpidemiologicoController = new casoEpidemiologicoController();
-	 	$casoEpidemiologicoController->deleteCasoEpidemiologicoController($_POST);
 	} else { 
-		
+		/*
 		session_start(["name"=> "dptoEpidemi"]);
 		session_unset();
 		session_destroy();
 		header("Location: ".SERVERURL."login/");
+		*/
 	}
